Update dest -> target in config.json

+++ entrypoint.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
 
 mkdir -p /etc/xray
-cp /config.json /etc/xray/config.json
+if [ ! -f /etc/xray/config.json ]; then
+    cp /config.json /etc/xray/config.json
+fi
 cp /url.txt /etc/xray/url.txt
 xray --config /etc/xray/config.json
